+++ src/globals.ts
@@ -4,7 +4,7 @@ export const MIN_SEARCH_LENGTH = 3;
 export const PAUSE = 100;
 export const TEXT_SEARCHING = "Searching...";
 export const TEXT_NO_RESULTS = "No results found";
-export const CLEAR_TIMEOUT = 100;
+export const CLEAR_TIMEOUT = 50;
 
 export function isNil(value: string) {
     return typeof value === "undefined" || value === null;